chore: Update action.yaml to include build-image and helm-install steps
This commit modifies the `action.yaml` file to include two new steps: `build-image` and `helm-install`. The `build-image` step is used to build a Docker image using the `docker/build-push-action@v5` action, allowing for the creation of a multi-platform image. The `helm-install` step is used to install a Helm chart using the `azure/setup-helm@v4.2.0` action, enabling the deployment of the application to a Kubernetes cluster. These changes enhance the deployment workflow by adding support for building Docker images and deploying Helm charts.