chore: Add conditional statement for deploy-infra step
This commit adds a conditional statement to the `action.yaml` file for the `deploy-infra` step. The conditional statement checks the value of the `deploy-infra` input variable and only executes the step if the input value is set to `true`. This change enhances the flexibility of the deployment workflow by allowing the `deploy-infra` step to be enabled or disabled based on user input.