chore: Add conditional statement for build-image step
This commit adds a conditional statement to the `action.yaml` file for the `build-image` step. The conditional statement checks the value of the `build-image` input variable and only executes the step if the input value is set to `true`. This change enhances the flexibility of the deployment workflow by allowing the `build-image` step to be enabled or disabled based on user input.