chore: Add conditional statement for helm-install step
This commit adds a conditional statement to the `action.yaml` file for the `helm-install` step. The conditional statement checks the value of the `helm-install` input variable and only executes the step if the input value is set to `true`. This change enhances the flexibility of the deployment workflow by allowing the `helm-install` step to be enabled or disabled based on user input.