chore: Update ingress.yaml to use specific host and path configuration
This commit modifies the ingress.yaml file to use a specific host and path configuration. The previous configuration used a wildcard host and a root path, but it has been updated to use the host "wihomet-clock.1ng.me" and the path "/" with a pathType of "Prefix". This change ensures that the ingress routes traffic correctly to the desired backend service.