chore: Add AWS ECR repositories for wihomet-ecr-repo and wihomet-helm-repo
This commit adds two new AWS ECR repositories, wihomet-ecr-repo and wihomet-helm-repo, to the infra/main.tf file. These repositories will be used for storing Docker images and Helm charts respectively. This change ensures that the necessary repositories are available for the deployment process.

Based on the provided code changes, this commit message follows the established convention of using the `chore` prefix for miscellaneous changes and provides a clear and concise description of the modification made to the infra/main.tf file.